pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gcp-mcp-server"
version = "0.1.0"
description = "MCP server for Google Cloud Platform"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "asklokesh", email = "49921737+asklokesh@users.noreply.github.com" }
]
dependencies = [
    "mcp>=1.27,<2",
    "pydantic>=2.12",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.0",
    "anyio>=4.9",
    "google-auth>=2.36",
    "google-cloud-storage>=2.19",
    "google-cloud-compute>=1.23",
    "google-cloud-bigquery>=3.27",
    "google-cloud-pubsub>=2.27",
    "google-cloud-logging>=3.12",
    "google-cloud-resource-manager>=1.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
gcp-mcp = "gcp_mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gcp_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]
